+++ lib/whois/record/parser/whois.registrypro.pro.rb
@@ -14,25 +14,23 @@ module Whois
   class Record
     class Parser
 
-      #
-      # = whois.registrypro.pro parser
-      #
       # Parser for the whois.registrypro.pro server.
       #
-      # NOTE: This parser is just a stub and provides only a few basic methods
-      # to check for domain availability and get domain status.
-      # Please consider to contribute implementing missing methods.
-      # See WhoisNicIt parser for an explanation of all available methods
-      # and examples.
+      # @note This parser is just a stub and provides only a few basic methods
+      #   to check for domain availability and get domain status.
+      #   Please consider to contribute implementing missing methods.
+      # 
+      # @see Whois::Record::Parser::Example
+      #   The Example parser for the list of all available methods.
       #
       class WhoisRegistryproPro < Base
 
         property_supported :status do
-          content_for_scanner.scan(/Status:(.*?)\n/).flatten
+          content_for_scanner.scan(/Status:(.+?)\n/).flatten
         end
 
         property_supported :available? do
-          !!(content_for_scanner =~ /No match/)
+          !!(content_for_scanner.strip == 'NOT FOUND')
         end
 
         property_supported :registered? do
@@ -41,26 +39,26 @@ module Whois
 
 
         property_supported :created_on do
-          if content_for_scanner =~ /Created On:(.*)\n/
+          if content_for_scanner =~ /Created On:(.+)\n/
             Time.parse($1)
           end
         end
 
         property_supported :updated_on do
-          if content_for_scanner =~ /Last Updated On:(.*)\n/
+          if content_for_scanner =~ /Last Updated On:(.+)\n/
             Time.parse($1)
           end
         end
 
         property_supported :expires_on do
-          if content_for_scanner =~ /Expiration Date:(.*)\n/
+          if content_for_scanner =~ /Expiration Date:(.+)\n/
             Time.parse($1)
           end
         end
 
 
         property_supported :nameservers do
-          content_for_scanner.scan(/Name Server:(.+)\n/).flatten.map do |name|
+          content_for_scanner.scan(/Name Server:([^\s]+)\n/).flatten.map do |name|
             Record::Nameserver.new(:name => name.downcase)
           end
         end